requirements.txt
anthropic>=0.25.0
tweepy>=4.14.0
Pillow>=10.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.35.0
pandas>=2.0.0
PyNaCl>=1.5.0
